Tests fixed
Former-commit-id: b8e2f9f86cf1fe65d8b5fce49741f2af553c5962
Former-commit-id: 8ba8872c84ef10a243e6698ec6e003bbd446cb58

+++ libs/app/alias_test.go
@@ -252,7 +252,7 @@ func Test_Alias(t *testing.T) {
 			})
 			Convey(`Should call list and respond with template`, func() {
 				app.App.Run([]string{"[app]", "list", "3", "5", "tag1"})
-				So(a.ListCalledWith, should.Resemble, []interface{}{"", int32(3), int32(5), []string{"tag1"}})
+				So(a.ListCalledWith, should.Resemble, []interface{}{"", int64(3), int64(5), []string{"tag1"}})
 				So(w.RenderCalledWith, should.Resemble, []interface{}{"alias:list", &models.AliasList{}})
 			})
 		})
