+++ libs/models/user.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	Id      int32
+	Id      int64
 	Username string
 	Email 	string
 	Token    string
